refactor: Extract the Flight server into its own module
